requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
Django==4.0.1
django-bootstrap-v5==1.0.8
django-location-field==2.1.0
importlib-metadata==3.10.1
Pillow==9.0.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
zipp==3.7.0
